requirements.txt
django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
requests==2.32.5
python-dotenv==1.2.1
supabase==2.24.0
